replaced NSUInteger with 'unsigned long' as the former does not exist <10.5
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@3761 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Output/gemcocoawindow.mm
@@ -88,7 +88,7 @@ struct gemcocoawindow :: PIMPL {
   NSWindow*window;
 //  WindowResponder*delegate;
   float titleBarHeight, menuBarHeight;
-  NSUInteger modifierFlags;
+  unsigned long modifierFlags;
 
   PIMPL(void) :
 	view(NULL),
@@ -233,10 +233,10 @@ case(NSKeyUp):
 break;
 case(NSFlagsChanged):
   do {
-   NSUInteger newflags = [e modifierFlags];
-   NSUInteger oldflags = m_pimpl->modifierFlags;
+   unsigned long newflags = [e modifierFlags];
+   unsigned long oldflags = m_pimpl->modifierFlags;
   if(newflags != oldflags) {
-   NSUInteger modified = newflags ^ oldflags;
+   unsigned long modified = newflags ^ oldflags;
    m_pimpl->modifierFlags = newflags;
 #define MODFLAGS2KEY(mask, name) if(modified & mask) key(name, [e keyCode], static_cast<bool>(mask & newflags))
    MODFLAGS2KEY(NSAlphaShiftKeyMask, "AlphaShift");
